requirements.txt
dnspython
motor
psutil
kurigram
python-dotenv
telethon
TgCrypto
